Remove Forced Error reporting
Although we could argue about whether error reporting should or shouldn't be set within a PHP File, the theme I submitted on Themeforest was rejected when bundled with Redux that includes error_reporting(). I had to manually remove it. I would love to keep up with the updates without doing merging myself all the time. So it would be nice if Redux just complied with Themeforest rules :)