Added new field type "section"
With this field type you can now add additional <h3> headlines between
options on the tabs pages.  It is also possible to indent the options
after a section element a bit to the right to make them look like a
group of options.

It is now possible to get a clearer structure in the tabs content when
you have a lot of options on one page/tab.