Fixed spelling error, again.
Gentlemen, I've address this before.  One does not 'loose' settings.  One 'loses' them.  Loose is the opposite of tight, so 'loose settings' means they are run around loose, or the settings in general are not tight.  Here endeth the grammar lesson.  We don't want our customers to think we cannot spell properly.  :)