Fix: PHP Fatal Error if "redux" folder not created
There is a Fatal Php Error which you can see in this screenshot http://jmp.sh/or4Wc5t and it is occurring if the folder "redux" is not created in wp-content  → uploads. I simply added functionality to create the folder and refresh the page and sorry this is my first contribution so forgive me for any mistakes.